Make sure replies cases if not found abort 404 exception

+++ modules/Forum/Application/UseCases/Reply/DeleteReplyUseCase.php
@@ -17,6 +17,7 @@ class DeleteReplyUseCase
     {
         $reply = $this->findReplyRepository->handle($id);
 
+        abort_if(!$reply, 404);
         Gate::authorize('delete', $reply);
 
         $reply?->delete();
